requirements.txt
streamlit>=1.30.0
pdfplumber
requests
beautifulsoup4
plotly
python-dotenv
urllib3
